INTERMEDIATE! Found a race condition in WDT. Finished PWM outputs.
The next items to do are to move the PWM generator to Timer1, and have it output PPM.
No more battery status PWM on the OCR1A/B Line to the blue LED
This will leave Timer 2 be used for packet timing.
Currently, Timer0 is sort of doing the packet timing.
May use one for expected packet delivery time, other left free.