Found Critical Bug in LRS Freq Calculation. Still Corrupt on Tx
Found the bug when using the current modifications to transmit along the offset range. Got 436 MHz when wanted 433 MHz.
The LRS Channel Set method in the rfm22b library was not casting operands correctly.
On the 23BP, the severe issue still remains that the RFM is corrupted after starting a TX.


KatanaLRS:
Main Loop modified for transmitting along the offset.
Was considering enabling the Low Battery Detector, as the Saleae scope is showing what may be brownout behavior:
	The last thing commited before it goes unstable is ordering Tx Manual (0x07 <- 0x0A)
	200 ms later, the system checks and write for TUNE, but the state is 0x82 on 0x07, so the Reset bit is active and Tx is done.
	Just about every register interaction thereafter is flagging an error and showing bogous values.
When the system goes unstable after attempting a manual Tx, I use the 'J' console command to dump registers, looks like everything is getting nuked, not to default values, but total messed-up ones.
EMI/RFI?

RFM22 Library:
Corrected some loops to i==199 instead of the faulty 200 on assert for loops.
Note: Need to make a utility method called assert, it would do all this for loop stuff.
Obviously, changed the rfmSetLrsChannel method to work.
Changed GPIO2 to Preamble output
Back-calculated the IF Loop and Clock coefficients, everything checks out, bumped the 0x1C value to 0x2B to cover the middle of the band the FHSS jumps in.
Enabled the AFC
The FHSS settings array is corrected back to the DragonLink Specs:
	Receive PH with reversed bits
	3 Sync Words
	3 nibbles of Preamble length. Learned more on the length requirement, must be honest when declaring length.
	1 nibble of Sync detect, not sure if it helps.
	Two versions of Sync word, use 0x90 0x42 0xE1, the 0xD0 0x50 are legacy/weird

Signed-off-by: Stephen Carlson <stephen_carlson@byu.edu>