DL FHSS Following Rx w/ AFC Offset
This commit demonstrates the KatanaLRS following the DragonLink FHSS stream, but not at full sequence, but on missing most cycles and only catching the next.
I have included + and - in the UART jump table to manually increase and decrease the freqOffset variable.
In this test, I had to increase the freqOffset to +180 to catch the DL stream. All the frequencies jumpped to seem to be the same offset of 175 to 180.

KatanaLRS:
UART Rx Jump Table additions.
freqOffset now used to filter and store AFC offset. Declared static volatile for UART.
Redo of frequency formulas.
DL Channel is incremented on RFM_INT, manualFreq+freqOffset sent to RFM.
As mentioned last commit, changed to RFM22B for the device under test.

Signed-off-by: Stephen Carlson <stephen_carlson@byu.edu>

Ammendment: Added GPIO2 System Duty Indication, reduced packet size, now blazes through the list full-speed until it hicups.

Signed-off-by: Stephen Carlson <stephen_carlson@byu.edu>