Entry point for function pointer state machine.
This version is the entry point for switching over to a function-pointer-based state machine.
This architecture is suppose to be more efficient and easy to maintain.
However, I don't feel like losing the ability to do printf's of the current system state.
And, I don't know, I'm just not proficcient in this new way of coding, and don't feel like breaking the entire project.
I'll keep the case statement for the time being.

Signed-off-by: Stephen Carlson <stephen_carlson@byu.edu>